d_a_npc_uk: match getters + chkGameStart (12 fns 100%)

+++ include/d/actor/d_a_npc_uk.h
@@ -69,7 +69,7 @@ public:
 
     u8 getType();
     u8 getPath();
-    u8 getShapeType();
+    int getShapeType();
     u16 getCaughtFlag();
     u16 getFoundFlag();
     u32 getFirstTalk();
